refactor : edit scan location each data(jpa & mongoDB) #37
