+++ Trim-Api/src/main/java/trim/api/domains/survey/service/GetSpecificSurveyUseCase.java
@@ -2,8 +2,8 @@ package trim.api.domains.survey.service;
 
 import lombok.RequiredArgsConstructor;
 import trim.common.annotation.UseCase;
-import trim.domains.survey.business.adaptor.SurveyAdaptor;
-import trim.domains.survey.dao.entity.Survey;
+import trim.doc.survey.business.adaptor.SurveyAdaptor;
+import trim.doc.survey.dao.entity.Survey;
 
 @UseCase
 @RequiredArgsConstructor
