+++ Trim-Api/src/main/java/trim/api/domains/survey/service/RegisterSurveyUseCase.java
@@ -1,12 +1,10 @@
 package trim.api.domains.survey.service;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.transaction.annotation.Transactional;
 import trim.api.domains.survey.vo.request.SurveyRequest;
 import trim.common.annotation.UseCase;
-import trim.domains.member.business.adaptor.MemberAdaptor;
-import trim.domains.survey.business.service.SurveyDomainService;
-import trim.domains.survey.dao.entity.Survey;
+import trim.doc.survey.business.service.SurveyDomainService;
+import trim.doc.survey.dao.entity.Survey;
 
 @UseCase
 @RequiredArgsConstructor
