+++ Trim-Api/src/main/java/trim/api/domains/survey/vo/request/SurveyRequest.java
@@ -1,10 +1,9 @@
 package trim.api.domains.survey.vo.request;
 
-import jakarta.persistence.Id;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import trim.domains.survey.dao.entity.SurveyQuestion;
+import trim.doc.survey.dao.entity.SurveyQuestion;
 
 import java.util.List;
