+++ Trim-Domain/src/main/java/trim/config/JpaConfig.java
@@ -4,7 +4,7 @@ import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import trim.DomainPackageLocation;
+import trim.domains.DomainPackageLocation;
 
 @Configuration
 @EnableJpaAuditing
