+++ Trim-Domain/src/main/java/trim/config/MongoDBConfig.java
@@ -9,20 +9,10 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.config.EnableMongoAuditing;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+import trim.doc.DocumentPackageLocation;
 
 @Configuration
 @EnableMongoAuditing
+@EnableMongoRepositories(basePackageClasses = DocumentPackageLocation.class)
 public class MongoDBConfig {
-
-//    private final MongoProperties mongoProperties;
-//
-//    @Bean
-//    public MongoClient mongoClient() {
-//        return MongoClients.create(mongoProperties.getUri());
-//    }
-//
-//    @Bean
-//    public MongoTemplate mongoTemplate(MongoClient mongoClient) {
-//        return new MongoTemplate(mongoClient, mongoProperties.getDatabase());
-//    }
 }
