+++ Trim-Domain/src/main/java/trim/doc/DocumentPackageLocation.java
@@ -1,0 +1,4 @@
+package trim.doc;
+
+public interface DocumentPackageLocation {
+}
