+++ Trim-Domain/src/main/java/trim/doc/survey/business/service/SurveyDomainService.java
@@ -1,0 +1,8 @@
+package trim.doc.survey.business.service;
+
+import trim.doc.survey.dao.entity.Survey;
+
+public interface SurveyDomainService {
+
+    Survey createSurvey(Survey survey);
+}
