+++ Trim-Domain/src/main/java/trim/doc/survey/dao/repository/SurveyAnswerRepository.java
@@ -1,11 +1,10 @@
-package trim.domains.survey.dao.repository;
+package trim.doc.survey.dao.repository;
 
 import org.springframework.data.mongodb.repository.Aggregation;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
-import org.springframework.stereotype.Repository;
-import trim.domains.survey.dao.entity.AnswerStatistics;
-import trim.domains.survey.dao.entity.SurveyAnswer;
+import trim.doc.survey.dao.entity.AnswerStatistics;
+import trim.doc.survey.dao.entity.SurveyAnswer;
 
 import java.util.List;
