--- Trim-Domain/src/main/java/trim/domains/survey/business/service/SurveyDomainService.java
@@ -1,8 +1,0 @@
-package trim.domains.survey.business.service;
-
-import trim.domains.survey.dao.entity.Survey;
-
-public interface SurveyDomainService {
-
-    Survey createSurvey(Survey survey);
-}
